requirements.txt
asgiref==3.2.10
Django==3.1.2
pytz==2020.1
sqlparse==0.4.1
psycopg2-binary==2.8.6
dj-database-url==0.5.0
django-extensions==3.0.9
ipython==7.18.1
djangorestframework==3.12.1
Markdown==3.3.2
django-filter==2.4.0
drf-yasg==1.20.0
djangorestframework-simplejwt==4.4.0
whitenoise==5.2.0
django-cors-headers==3.5.0
model-bakery==1.2.0
django-widget-tweaks==1.4.8
Pillow==8.0.1
django-heroku==0.3.1
gunicorn==20.0.4
django-storages==1.10.1
boto3==1.16.11
